requirements.txt
click>=8.1.2
Flask>=2.1.1
importlib-metadata>=4.11.3
itsdangerous>=2.1.2
Jinja2>=3.1.1
MarkupSafe>=2.1.1
pymongo>=4.1.1
Werkzeug>=2.1.1
zipp>=3.8.0
python-dotenv>=0.20.0
aniso8601>=9.0.1
certifi>=2021.10.8
charset-normalizer>=2.0.12
click>=8.1.2
Flask>=2.1.1
Flask-RESTful>=0.3.9
googlemaps>=4.6.0
idna>=3.3
importlib-metadata>=4.11.3
itsdangerous>=2.1.2
Jinja2>=3.1.1
MarkupSafe>=2.1.1
pymongo>=4.1.1
python-dotenv>=0.20.0
pytz>=2022.1
requests>=2.27.1
six>=1.16.0
urllib3>=1.26.9
Werkzeug>=2.1.1
zipp>=3.8.0